Moved TestActivities into a separate file.
