+++ temporal-sdk/src/test/java/io/temporal/workflow/ActivityApplicationFailureNonRetryableTest.java
@@ -34,8 +34,8 @@ import org.junit.Test;
 
 public class ActivityApplicationFailureNonRetryableTest {
 
-  private final WorkflowTest.TestActivitiesImpl activitiesImpl =
-      new WorkflowTest.TestActivitiesImpl(null);
+  private final TestActivities.TestActivitiesImpl activitiesImpl =
+      new TestActivities.TestActivitiesImpl(null);
 
   @Rule
   public TestWorkflowRule testWorkflowRule =
@@ -72,7 +72,7 @@ public class ActivityApplicationFailureNonRetryableTest {
   public static class TestActivityApplicationFailureNonRetryable
       implements WorkflowTest.TestWorkflow1 {
 
-    private WorkflowTest.TestActivities activities;
+    private TestActivities activities;
 
     @Override
     public String execute(String taskQueue) {
@@ -87,7 +87,7 @@ public class ActivityApplicationFailureNonRetryableTest {
                       .setDoNotRetry(IOException.class.getName())
                       .build())
               .build();
-      activities = Workflow.newActivityStub(WorkflowTest.TestActivities.class, options);
+      activities = Workflow.newActivityStub(TestActivities.class, options);
       activities.throwIO();
       return "ignored";
     }
