+++ temporal-sdk/src/test/java/io/temporal/workflow/ActivityApplicationNoSpecifiedRetryTest.java
@@ -31,8 +31,8 @@ import org.junit.Test;
 
 public class ActivityApplicationNoSpecifiedRetryTest {
 
-  private final WorkflowTest.TestActivitiesImpl activitiesImpl =
-      new WorkflowTest.TestActivitiesImpl(null);
+  private final TestActivities.TestActivitiesImpl activitiesImpl =
+      new TestActivities.TestActivitiesImpl(null);
 
   @Rule
   public TestWorkflowRule testWorkflowRule =
@@ -69,7 +69,7 @@ public class ActivityApplicationNoSpecifiedRetryTest {
   public static class TestActivityApplicationNoSpecifiedRetry
       implements WorkflowTest.TestWorkflow1 {
 
-    private WorkflowTest.TestActivities activities;
+    private TestActivities activities;
 
     @Override
     public String execute(String taskQueue) {
@@ -79,7 +79,7 @@ public class ActivityApplicationNoSpecifiedRetryTest {
               .setScheduleToCloseTimeout(Duration.ofSeconds(200))
               .setStartToCloseTimeout(Duration.ofSeconds(1))
               .build();
-      activities = Workflow.newActivityStub(WorkflowTest.TestActivities.class, options);
+      activities = Workflow.newActivityStub(TestActivities.class, options);
       activities.throwApplicationFailureThreeTimes();
       return "ignored";
     }
