+++ temporal-sdk/src/test/java/io/temporal/workflow/ActivityRetryOnTimeoutTest.java
@@ -32,8 +32,8 @@ import org.junit.Test;
 import org.junit.rules.TestName;
 
 public class ActivityRetryOnTimeoutTest {
-  private final WorkflowTest.TestActivitiesImpl activitiesImpl =
-      new WorkflowTest.TestActivitiesImpl(null);
+  private final TestActivities.TestActivitiesImpl activitiesImpl =
+      new TestActivities.TestActivitiesImpl(null);
 
   @Rule public TestName testName = new TestName();
 
@@ -89,8 +89,7 @@ public class ActivityRetryOnTimeoutTest {
                       .setDoNotRetry(AssertionError.class.getName())
                       .build())
               .build();
-      WorkflowTest.TestActivities activities =
-          Workflow.newActivityStub(WorkflowTest.TestActivities.class, options);
+      TestActivities activities = Workflow.newActivityStub(TestActivities.class, options);
       long start = Workflow.currentTimeMillis();
       try {
         activities.neverComplete(); // should timeout as scheduleToClose is 1 second
