+++ temporal-sdk/src/test/java/io/temporal/workflow/AsyncUntypedActivityTest.java
@@ -28,8 +28,8 @@ import org.junit.Test;
 
 public class AsyncUntypedActivityTest {
 
-  private final WorkflowTest.TestActivitiesImpl activitiesImpl =
-      new WorkflowTest.TestActivitiesImpl(null);
+  private final TestActivities.TestActivitiesImpl activitiesImpl =
+      new TestActivities.TestActivitiesImpl(null);
 
   @Rule
   public TestWorkflowRule testWorkflowRule =
