+++ temporal-sdk/src/test/java/io/temporal/workflow/UntypedActivityRetryTest.java
@@ -33,8 +33,8 @@ import org.junit.Test;
 
 public class UntypedActivityRetryTest {
 
-  private final WorkflowTest.TestActivitiesImpl activitiesImpl =
-      new WorkflowTest.TestActivitiesImpl(null);
+  private final TestActivities.TestActivitiesImpl activitiesImpl =
+      new TestActivities.TestActivitiesImpl(null);
 
   @Rule
   public TestWorkflowRule testWorkflowRule =
